test: switch loader tests to react testing library
